Momentum Scrolling works
The momentum scrolling is now integrated into the Chart.coffee script.
In init() a @momentum map is created that contains several config
properties as well as a couple of working variables.
> fallOff: The 'friction coefficient' that slows the scrolling down
> stopThreshold: Allows a cancel of the expensive calculations if the
speed is too low to matter.
> maxDuration: Allows a forced cancel of the animation after this amount
of ms.

Do not touch the other properties unless you are modifying the
implementation itself!